SWCForecastBase

Documentation for SWCForecastBase.

SWCForecastBase.Cache

SWCForecastBase.ConfigAccumulate

SWCForecastBase.ConfigPreprocess

SWCForecastBase.ConfigSeriesToSupervised

SWCForecastBase.DataInterval

SWCForecastBase.DataRatio

SWCForecastBase.DataRatio

SWCForecastBase.PrepareTable

SWCForecastBase.PrepareTable

SWCForecastBase.SeriesToSupervised

SWCForecastBase.TrainTestState

MakieCore.convert_arguments

MakieCore.heatmap!

SWCForecastBase.PrepareTableDefault

SWCForecastBase._check

SWCForecastBase._series2supervised

SWCForecastBase.accu_unit_suffix_function

SWCForecastBase.accu_unit_suffix_function

SWCForecastBase.addcol_accumulation!

SWCForecastBase.cccount

SWCForecastBase.chknnm

SWCForecastBase.convert_types

SWCForecastBase.dataoverview

SWCForecastBase.dataratio

SWCForecastBase.diffsstable!

SWCForecastBase.fstree

SWCForecastBase.imputeinterp!

SWCForecastBase.imputemean!

SWCForecastBase.islnan

SWCForecastBase.islnan

SWCForecastBase.isnnm

SWCForecastBase.isoutofrange

SWCForecastBase.manytrees

SWCForecastBase.movingaverage

SWCForecastBase.narrow_types!

SWCForecastBase.precipmax!

SWCForecastBase.preparetable!

SWCForecastBase.preparetable!

SWCForecastBase.preparetable!

SWCForecastBase.removeunreasonables!

SWCForecastBase.save

SWCForecastBase.save

SWCForecastBase.series2supervised

SWCForecastBase.take_hour_last

SWCForecastBase.test!

SWCForecastBase.train!

SWCForecastBase.traintest!

SWCForecastBase.transform_datetime!

SWCForecastBase.twofstree

SWCForecastBase.unithour

SWCForecastBase.Cache — Type

Field

prepare::Prepare
train::Train
test::Test

See also TrainTestState.

SWCForecastBase.ConfigAccumulate — Type

ConfigAccumulate is the configuration for generating derived variables as the new columns, by accumulating each variable in variables for every interval in each intervals.

Use keyword arguments to construct the object.

Example

ConfigAccumulate(; variables = Cols(:precipitation_max),
                                intervals = [1, 12, 24, 48, 36],
                                unit = "hr"
        )

variables: Cols column selector for selecting columns (variables) to be derived.

intervals: the window length for accumulating an variable.

unit: unit of intervals as appended string of the new column, e.g., day.

unit can be an arbitrary function, for example:

unithour

See accu_unit_suffix_function and addcol_accumulation!.

SWCForecastBase.ConfigPreprocess — Type

ConfigPreprocess is the configuration for a table of time series input/target features.

ConfigPreprocess controls the primary feature selection and how the data being preprocessed before training.

Use keyword arguments to construct the object.

Example

ConfigPreprocess(;
        timeargs = Cols(:year, :month, :day, :hour),
        input_features  = Cols(r"air_temp", r"humidity", r"pressure", r"wind_speed", r"precipitation"),
        target_features = Cols(r"soil_water_content"),
        preprocessing   = [take_hour_last, removeunreasonables!, imputeinterp!, disallowmissing!, precipmax!],
        )

PrepareTable with ConfigPreprocess creates a new column :datetime according to timeargs. See preparetable!.

SWCForecastBase.ConfigSeriesToSupervised — Type

ConfigSeriesToSupervised controls how the data being shifted; it is for preparing the data for supervised model training.

Use keyword arguments to construct the object.

Example

using SWCForecastBase: ConfigSeriesToSupervised
config1 = ConfigSeriesToSupervised(;
        shift_x         = [0, -6, -12, -24, -36, -48, -60, -72],
        shift_y         = [1],
        shift_t         = [1]
        )

config2 = ConfigSeriesToSupervised(;
        shift_x         = [0, -6, -12, -24, -36, -48, -60, -72],
        shift_y         = [1],
        ) # by default, `shift_t` is fixed to the time of target feature (`shift_y`) if not assigned manually.

config1 == config2

# output

true

which is equivalent to

shift_x: the time shift for the data using as the input features

shift_y: the time shift for the data using as the target features

SWCForecastBase.DataInterval — Type

struct DataInterval
    from
    until
    identifier
end

for DataRatio.

SWCForecastBase.DataRatio — Type

struct DataRatio
    table::DataFrame
    dataintervals::StructArray
end

An DataRatio object that stores the results of dataratio.

SWCForecastBase.DataRatio — Method

DataRatio constructor accept exactly the same arguments as dataratio.

SWCForecastBase.PrepareTable — Type

Constructor

PrepareTable(table) = new(table, PrepareTableConfig[])

Field

table::DataFrame
configs::Vector{<:PrepareTableConfig}
status::Union{TrainTestState, Nothing}
supervised_tables::Union{SeriesToSupervised, Nothing}
cache::Cache

Example

    PrepareTable(df::DataFrame, ConfigPreprocess(), ConfigAccumulate(), ConfigSeriesToSupervised())

is equivalently

    DefaultPrepareTable(df)

Another example

PrepareTable(df,
    ConfigPreprocess(;target_features=Cols(r"soil_water_content_10cm")),
    ConfigAccumulate(),
    ConfigSeriesToSupervised(; shift_x=[0, -2])
)

See also DefaultPrepareTable.

SWCForecastBase.PrepareTable — Method

Given a table::DataFrame and PTCs::PrepareTableConfig..., PrepareTable runs preparetable!(_, PTC::PrepareTableConfig) for PTC in PTCs in @chain.

Example

    PrepareTable(df::DataFrame, ConfigPreprocess(), ConfigSeriesToSupervised())

SWCForecastBase.SeriesToSupervised — Type

T is the datetime of Y.

SWCForecastBase.TrainTestState — Type

TrainTestState is an abstract type of Prepare, Train, and Test, which indicate the current status and be in the latest cache of the PrepareTable.

Status in the workflow

 `PT.status`:
        `nothing`         `Prepare`d
 ╠═══════════════════════╬════════════════════╬...

 PT::PrepareTable
  ┃
  ┗━━━━ preparetable! ━━━┳━━━━━━▶ preparetable!
         ▲               ┃        that creates
         ┗━━━━━━━━━━━━━━━┛    `T.supervised_tables`
       (preprocessing using                   ┃
        different configs)                    ┃
                                              ▼
                           ┏━━ test! ◀━━ train!
     (change parameters to ┃          ▲       ┃
     train or tested again)┗━━━━━━━━━━┻━━━━━━━┛

                         ..═══════════╬═════════..
                          .`Test`ed and `Train`ed

Field

args::NamedTuple

See also Cache.

MakieCore.convert_arguments — Method

convert_arguments(DR::DataRatio) returns xs, ys, vs for heatmap!(ax, xs, ys, vs,...). This is intended to extend Makie convert_arguments methods for DataRatio, but it does not work with Makie.heatmap. Checkout git branch try-convert_arguments for more details.

Example

using Dates, SWCExampleDatasets, CairoMakie
ari0 = SWCExampleDatasets.dataset("NCUWiseLab", "ARI_G2F820_example")
DR = DataRatio(ari0, Month(1), SWCForecastBase.islnan)
DR |> convert_arguments |> x -> heatmap(x...)

MakieCore.heatmap! — Method

Example

using CairoMakie, SWCForecastBase
f = Figure(; resolution=(800,600))
ax = Axis(f[1,1])
hmap = heatmap!(ax, DR::DataRatio; colormap = "diverging_rainbow_bgymr_45_85_c67_n256")
Colorbar(f[1, 2], hmap, label = "missing data rate")

See DataRatio and SWCForecastBase.convert_arguments.

SWCForecastBase.PrepareTableDefault — Method

Default data processing:

PrepareTableDefault(df::DataFrame) = PrepareTable(df, ConfigPreprocess(), ConfigAccumulate(), ConfigSeriesToSupervised())

SWCForecastBase._check — Method

_check(PT::PrepareTable, ::Nothing) check when the status is nothing.

SWCForecastBase._series2supervised — Method

series2supervised(data, range_shift, range_out)

Tansform a time series dataset into a supervised learning dataset.

The features will always be suffixed by an addtional time shift tag "t-i". Also see `splittimetag()andformattime_tag`.

References:

https://machinelearningmastery.com/convert-time-series-supervised-learning-problem-python/

SWCForecastBase.accu_unit_suffix_function — Method

accu_unit_suffix_function returns a suffix function sfx(i) for generates the keys for apd = Dict(sfx.(PTC.intervals) .=> PTC.intervals) that is going to be fed into addcol_accumulation!, where PTC::ConfigAccumulate.

accu_unit_suffix_function(PTCunit::AbstractString) returns i -> "$i$(PTCunit)".

This returned function is essential for generate readable tags appended as the new column names when using PrepareTable, PrepareTableDefault, or preparetable! with the configuration ConfigAccumulate.

SWCForecastBase.accu_unit_suffix_function — Method

If input is a function, it simply returns its argument: accu_unit_suffix_function(PTCunit::Function) = PTCunit.

This is for the case for customized suffix-generating function.

See also ConfigAccumulate.

SWCForecastBase.addcol_accumulation! — Method

Add columns that are derived by accumulating corresponding variables using movingaverage.

Example

all_precipstr = names(df, r"precipitation")

apd = Dict( # time intervals to accumulates precipitation
    "1hour" => 6,
    "12hour" => 6*12,
    "1day" => 6*24,
    "2day" => 6*24*2,
    "3day" => 6*24*3
)

addcol_accumulation!(df, all_precipstr, apd)

See also: movingaverage.

SWCForecastBase.cccount — Method

Of a time series ts, cccount(ts) calculate by default the cumulative counts of elements that approximates zero consecutively.

SWCForecastBase.chknnm — Method

chknnm(df) check if DataFrame df contains missing values or NaN. Use this before input df into machine.

SWCForecastBase.convert_types — Method

convert_types(df, column_names_types)

Converts the element type of each column to a user-specified type.

Arguments

df: Dataframe for which you want to convert the eltype of each column

column_names_types: Column names and target types. The type of column_names_types should be able to be unpacked into column names and target types in a for loop.

References

https://discourse.julialang.org/t/how-to-change-field-names-and-types-of-a-dataframe/43991/11

SWCForecastBase.dataoverview — Method

dataoverview(df::DataFrame, xtype::DataType; gridsize = gridsize, set_left_edge=leftfn, set_right_edge=rightfn) plot the overview of data in DataFrame df. xtype specify the only column type (e.g., DateTime) for x-axis.

Example

using Dates, DataFrames, OkMakieToolkits, Makie
a = randn(100)
a[1:20] .= NaN
b = Vector{Union{Missing, Float64}}(undef, 100)
b[1:70] .= randn(70)
b[71:90] .= NaN
b[91:100] .= missing
table_nan = DataFrame(:a => [1,2,NaN], :b => [missing,missing,5], :dt => collect(DateTime("2022-01-01T00:00:00"):Day(1):DateTime("2022-01-03T00:00:00")))
xs, ys, vs, xtick_label, ytick_label = dataoverview(table_nan, DateTime)

Makie.save("Fig_dataoverview.eps", f)

SWCForecastBase.dataratio — Method

dataratio(df::DataFrame, gridsize::DatePeriod, iswhat::Function;
                        groupbycol=:datetime)

Given data (df), and size of intervals in DatePeriod (gridsize), dataratio returns a combined dataframe of all Not(groupbycol) with interval_id where iswhat for each element in a column ByRow is true.

Also see DataRatio.

SWCForecastBase.diffsstable! — Method

Of a variable of name varnm, diffsstable!(X0::DataFrame, varnm, tshift) calculates the difference between the non-shifted (suffixed by "_t0") and time-shifted (e.g., "_t-6"), where the difference is the new column for the series-to-supervised table X0.

Example

    (X0,) = series2supervised(...)
    diffsstable!(X0, "precipitation_1hr", -6)

that creates a new column diff6_precipitation_1hr = X0[:, "precipitation_1hr_t0"] .- X0[:, "precipitation_1hr_t-6"].

SWCForecastBase.fstree — Method

Return a composite tree model with FeatureSelector.

SWCForecastBase.imputeinterp! — Method

Similar to imputemean!(df), imputeinterp!(df) interpolate literal nan values using linear interpolation, and extrapolate using Next Observation Carried Backward (NOCB) and Last Observation Carried Forward (LOCF).

See also imputemean!.

SWCForecastBase.imputemean! — Method

imputemean!(df) substitute literal nan values with the statistical means. If all missing for a column, value 999 is imputed.

Notice

You should be aware that imputemean! might does nothing without error message if the input is a view of dataframe (e.g., df[!, Not(:datetime)]).

See also islnan for literal nan.

SWCForecastBase.islnan — Method

Return true if it is literally not a number. For example, all(islnan.(["#VALUE!", "nan", "NaN", "Nan", nothing])) is true.

SWCForecastBase.islnan — Method

For x being the type other than the types listed above, islnan(x) falls back to isnnm(x).

See isnnm.

SWCForecastBase.isnnm — Method

Check if x is missing, nothing or NaN. Different from isnan, for x being either of Nothing, Missing, DateTime, and AbstractString, islnan(x) returns true for Nothing and Missing, and returns false for the rest.

The difference between islnan and isnnm is that, isnnm check only NaN for Not-a-Number. If you input something like "#VALUE!", "NaN", it returns false (NOT missing, nothing or NaN).

SWCForecastBase.isoutofrange — Method

Return true if value is out of the interval between l0 and l1. Noted that if value is missing, nothing, or literally nan (see islnan), it returns false (NOT out-of-range).

SWCForecastBase.manytrees — Method

A random forest using EnsembleModel

SWCForecastBase.movingaverage — Method

movingaverage(v, n) returns a vector of moving averaged v by every n.

SWCForecastBase.narrow_types! — Method

narrow_types!(df)

Narrows the eltype of each column to the type that actually exists in the each column of dataframe.

Arguments

df: Dataframe for which you want to narrow the eltype of each column

References

https://discourse.julialang.org/t/how-to-change-field-names-and-types-of-a-dataframe/43991/9

SWCForecastBase.precipmax! — Method

precipmax!(df::DataFrame) creates precipitation_max by maximize Cols(r"\Aprecipitation") ByRow. Noted that the original "precipitation*" columns were dropped; only :precipitation_max kept.

SWCForecastBase.preparetable! — Method

preparetable!(PT::PrepareTable, PTC::ConfigAccumulate) generates derived variables as new columns. See ConfigAccumulate.

SWCForecastBase.preparetable! — Method

preparetable!(PT::PrepareTable, PTC::ConfigPreprocess) generates datetime column by PTC.timeargs, sort! by :datetime, do PTC.preprocessing in @chain and check if the table is continuous in time.

Note

This method will raise essential error, unless that PTC::ConfigPreprocess is the first arg in args of PrepareTable(PT, args...) as it should be. This is intended since ConfigAccumulate or ConfigSeriesToSupervised is designed to work on a well preprocessed data, and it may give incorrect results without error (e.g., time tag mismatch) if preparetable! with ConfigPreprocess is not conducted before.

SWCForecastBase.preparetable! — Method

preparetable!(PT::PrepareTable, PTC::ConfigSeriesToSupervised) creates SeriesToSupervised as PT.supervised_tables for training and testing for supervised models.

SWCForecastBase.removeunreasonables! — Method

removeunreasonables!(df_all) convert all column-name specific unreasonable values to missing. Noted that missing, nothing and literal nan is not "unreasonable values".

Noted that removeunreasonables! will NOT deal with literally Not-a-Number value nor raising an error for any literally Not-a-Number value. See isoutofrange, islnan.

SWCForecastBase.save — Method

save(PT::PrepareTable, dir0) save cache in PT to the directory dir0.

Example

dir0 = "MyResults"
save(PT::PrepareTable, dir0)

SWCForecastBase.save — Method

Without specifying the parent folder dir0, a folder of name of randomstring will be created as the parent folder.

Example

save(PT::PrepareTable) # save results to /Tables_Fsd0w4/... for example

SWCForecastBase.series2supervised — Method

To transform a time series dataset into a supervised learning dataset

Example

A = randn(500,20)
df = DataFrame(A, :auto)
X0,y0 = series_to_supervised(df[:,1:end-1], df[:,end])
X1,y1 = series2supervised(
    df[:,1:end-1] => range(-6, -1; step=1),
    df[:,end] => range(0, 0; step=-1)
    )

NOTICE!

The input DataFrame (df) must have complete rows; that is, the corresponding time tag (it might be df.datetime for example) must be consecutive because df is converted to Matrix and shifted using lag.

This function filter the dataframe using completecases.

References:

https://machinelearningmastery.com/convert-time-series-supervised-learning-problem-python/

TODO: write test for series2supervised, by making sure the datetime shift is correct (e.g., "datetimet0" should always be 1 hour ahead of "datetimet-6" for a 10-minute sampling data).

SWCForecastBase.take_hour_last — Method

combine the df which were grouped by :hour taking only the last.

SWCForecastBase.test! — Method

Example

test!(PT::PrepareTable; test_after = :auto,  # default
    numpoints_test = 480)   # default

test_after: the DateTime or Date after which model prediction (model testing stage) starts. If :auto, test_after = now().

numpoints_test: the number of data points in the "future" to be tested.

SWCForecastBase.train! — Method

Training with PrepareTable.

Example

train!(PT::PrepareTable;
        train_before = :auto, # default
        model = manytrees(),    # default
        numpoints_train = 24*120 # default
    )

train_before: The DateTime or Date before which the data is used for training; if :auto, train_before = now().

model: The model; it can be any MLJ model.

numpoints_train: from train_before how many data points (number of rows) to be included for model training.

SWCForecastBase.traintest! — Method

traintest!(PT::PrepareTable; kwargs...) do train! then test!, accepts all keyword arguments that train! or test! could have.

SWCForecastBase.transform_datetime! — Method

transform_datetime!(df::DataFrame, groupbycol; list=[:year, :month, :day, :hour, :minute, :second]) attempts to convert the groupbycol = :datetime column from :year, :month, :day, :hour, :minute, :second.

Noted that once transformed, columns in list will be removed by default.

SWCForecastBase.twofstree — Method

Return a composite tree model with TWO FeatureSelector: selector_1 and selector_2.

SWCForecastBase.unithour — Method

unithour(x) is an exemplary suffix formatter.

SWCForecastBase.unithour(5)

# output

"5hr"

SWCForecastBase.unithour(24)

# output

"1d"

SWCForecastBase.unithour(36)

# output

"1d12hr"